Got rid of compiler warnings and added X,Y,Z FCVariables.

+++ Uintah/Modules/Visualization/TensorFieldExtractor.cc
@@ -113,7 +113,7 @@ void TensorFieldExtractor::setVars()
 
 
   // get all of the TensorField Variables
-  for( int i = 0; i < names.size(); i++ ){
+  for( int i = 0; i < (int)names.size(); i++ ){
     const TypeDescription *td = types[i];
     const TypeDescription *subtype = td->getSubType();
     //cerr << "\tVariable: " << names[i] << ", type " << td->getName() << "\n";
